api-back/requirements.txt
django
djangorestframework
django-rest-auth
django-allauth
mysqlclient
gunicorn
django-cors-headers
pdfkit